Conflicts:
	semi/.classpath
	semi/src/com/bbbox/lawyer/model/vo/Counsel.java
	semi/src/com/bbbox/lawyer/model/vo/LawReview.java
	semi/src/com/bbbox/lawyer/model/vo/Lawyer.java

+++ semi/WebContent/views/board/accidentBoardDetailView.jsp
@@ -8,11 +8,14 @@
 	Board b = (Board)request.getAttribute("board");
 	Accident ac = (Accident)request.getAttribute("accident");
 	ArrayList<Reply> rplist = (ArrayList<Reply>)request.getAttribute("rplist");
+	
 %>
 <!DOCTYPE html>
 <html>
 <head>
 <meta charset="UTF-8">
+<!-- font-awesome Copy Link Tag (아이콘 CDN) -->
+<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css" integrity="sha512-iecdLmaskl7CVkqkXNQ/ZH/XLlvWZOJyj7Yy7tcenmpD1ypASozpmT/E0iPtmFIB46ZmdtAc9eNBvH0H/ZpiBw==" crossorigin="anonymous" referrerpolicy="no-referrer" />
 <title>제보영상 게시글</title>
 <style>
         div{
@@ -136,30 +139,55 @@
         <div class="accidentBoardDetailBody">
             <div class="bodyLeft">
                 <div class="videoArea">
-                    <img src="" id="video"></img>
+                    <video src="<%=contextPath%>/resources/accident_board_file/video1.mp4" controls poster="" id="video"></video>
                 </div>
                 <div class="contentArea">
                     <div class="infoArea">
                         <div class="detail">
                             <div class="title">
                                 <span><%=b.getTitle()%></span>
-                                <span>#미해결</span>
-                                <span>#보험</span>
+                                <%if(ac.getSolve().equals("N")) {%>
+                                	<span>#미해결</span>
+                                <%}else{ %>
+                                	<span>#해결</span>
+                                <%} %>
+                                <span>#<%=ac.getInsuranceType()%></span>
                             </div>
                             <div class="info">
                                 <span>제보일 | <%=b.getCreateDate()%></span>
                                 <span>조회수 | <%=b.getCount() %></span>
+                                <span>신고수 | <%=b.getReportCount()%> </span>
+                                <%if(b.getReportCount()>10){ %>
+                                	<span><i class="fa-solid fa-diamond-exclamation" style="color: #e6d519;"></i></span>
+                                <%}else if(b.getReportCount()>20){ %>
+                                	<i class="fa-solid fa-diamond-exclamation" style="color: #f80d0d;"></i>
+                                <%} %>
                                 <br>
                                 <span>제보자 | <%=b.getBoardWriter()%></span>
                                 <span>지역 | <%=ac.getRegion()%></span>
                             </div>
-                            <%if(loginUser!=null) {%>
-	                            <div id="updateBtnArea" style="text-align:right;">
-	                            	<!-- 나중에 이미지로 바꿔서 onclick 이벤트 -->
-	                            	<input type="button" value="게시글 수정" onclick="location.href='<%=contextPath%>/update.ac?bno='+<%=b.getBoardNo()%>">
-	                            	<input type="button" value="게시글 삭제" onclick="location.href='<%=contextPath%>/delete.ac?bno='+<%=b.getBoardNo()%>">
-	                            </div>
+                            <%if(loginUser!=null){ %>
+	                            <%if(loginUser.getUserId().equals(b.getBoardWriter())||loginUser.getAdmin().equals("Y")) {%>
+	                        
+		                            <div id="updateBtnArea" style="text-align:right;">
+		                            	<!-- 나중에 이미지로 바꿔서 onclick 이벤트 -->
+		                            	<input type="button" value="게시글 수정" onclick="location.href='<%=contextPath%>/update.ac?bno='+<%=b.getBoardNo()%>">
+		                            	<input type="button" value="게시글 삭제" onclick="location.href='<%=contextPath%>/delete.ac?bno='+<%=b.getBoardNo()%>">
+		                            </div>
+	                            <%} %>
                             <%} %>
+                            <%if(loginUser!=null){ %>
+	                            <%if(loginUser.getLawyer().equals("Y")){ %>
+	                            	<div id="accidentReviewBtn">
+	                            		<form action="enroll.ar">
+	                            			<input type="hidden" name="oldFileChangeName" value="<%=b.getChangeName()%>">
+	                            			<input type="hidden" name="bno" value="<%=b.getBoardNo()%>">
+	                            			<input type="hidden" name="accNo" value="<%=ac.getAccNo()%>">
+	                            			<input type="submit" value="해결완료하려면 리뷰를 작성하세요">
+	                            		</form>
+	                            	</div>
+	                            <%} %>
+	                        <%} %>
                         </div>
                         <!-- 과실비율 -->
                         <div class="rate">
@@ -182,7 +210,7 @@
                 </div>
                 <!-- 댓글작성 영역 -->
                 
-                <div class="replyWriteArea">
+                <div class="replyWriteArea" >
                 <%if(loginUser!=null){ %>
                     <table>
                         <tr>
@@ -192,41 +220,97 @@
                             <td><%=loginUser.getUserId()%></td>
                         </tr>
                     </table>
-                    <textarea name="" id="" cols="70" rows="5" style="resize: none;" placeholder="댓글을 입력해주세요."></textarea>
-                    <input type="submit" value="댓글작성">
+                    <textarea name="" id="reply_input" cols="70" rows="5" style="resize: none;" placeholder="댓글을 입력해주세요."></textarea>
+                    <input type="button" id="reply_btn" value="댓글등록">
                 <%}else{ %>
                 	<table>
                         <tr>
                             <th>작성자</th>
                         </tr>
                         <tr>
-                            <td>-</td>
+                            <td>비회원</td>
                         </tr>
                     </table>
-                    <textarea name="" id="" cols="70" rows="5" style="resize: none;" placeholder="로그인한 유저만 이용 가능합니다." readonly></textarea>
-                    <input type="submit" value="댓글작성" disabled>
+                    <textarea name="" id="" cols="70" rows="5" style="resize: none;" placeholder="로그인한 유저만 이용 가능합니다." disabled></textarea>
+                    <input type="button" value="댓글작성" disabled>
                 <%} %>    
                 </div>
             </div>
             <div class="bodyRight">
-                <div class="replyViewArea">
-                    <%for(int i=0;i<rplist.size();i++){ %>
-	                    <div>
-	                        <table>
-	                            <tr>
-	                                <th><%=rplist.get(i).getRpWriter()%></th>
-	                            </tr>
-	                            <tr>
-	                                <td><%=rplist.get(i).getContent() %></td>
-	                            </tr>
-	                        </table>
-	                    </div>
-                    <%} %>
+                <div id="replyViewArea">
+                    <table>
+                    	
+                    </table>
+	                
                 </div>
-                <div class="recommendArea">?</div>
+                <div class="recommendArea">광고맨</div>
             </div>
         </div>
-       
     </div>
+	        <script>
+	        	//디테일 넘어왔을때 댓글 최신화
+		        $(function(){
+					viewRpList();
+				});
+	        
+	        	$(function(){
+	        		$("#reply_input").on("keydown",function(e){
+	        			//엔터키를 눌렀을때
+	        			if(e.keyCode ==13){
+	        				e.preventDefault(); //줄바꿈기능을 제거하고 
+	        				$("#reply_btn").click(); //버튼의 submit과 동일한기능 추가
+	        			}
+	        		});
+	        		
+	        		$("#reply_btn").on("click",function(){
+	        			<%if(loginUser !=null){%>
+	        				$.ajax({
+	        					url:"insertRp",
+	        					data:{
+	        						bno:<%=b.getBoardNo()%>,
+	        						content:$("#reply_input").val()
+	        					},
+	        					success:function(result){
+	        						if(result>0){
+	        							alert("댓글 등록 성공");
+	        							//완료되면 댓글창 비워주기
+	        							$("#reply_input").val("");
+	        							
+	        							//댓글 리스트 최신화
+	        							viewRpList();
+	        						}else{
+	        							alert("댓글 등록 실패");
+	        						}
+	        					},
+	        					error:function(){
+	        						alert("댓글 등록 통신 실패");
+	        					}
+	        				});
+	        			<%}%>
+	        		});
+	        	});
+	        	
+	        	function viewRpList(){
+	        		$.ajax({
+	        			url:"listRp",
+	        			data:{
+    						bno:<%=b.getBoardNo()%>
+    					},
+	        			success:function(list){
+	        				var str = "";
+	        				
+	        				for(var i in list){
+	        					str += "<tr><td>"+list[i].rpWriter+"</td>"
+								  +"<td style='text-align:left; padding-left: 5px;'>"+list[i].content+"</td>"
+								  +"<td>"+list[i].createDate+"</td></tr>";
+	        				}
+	        				$("#replyViewArea>table").html(str);
+	        			},
+	        			error:function(){
+	    					console.log("댓글리스트 조회 통신 실패");
+	    				}
+	        		});
+	        	}
+	        </script>
 </body>
 </html>
